--story=869998977 【流水线清理】自动清理逻辑优化：MR目标分支上没有触发过CI，且YML未删除时，不清理流水线执行记录

+++ src/backend/ci/ext/tencent/stream/biz-stream-tencent/src/main/kotlin/com/tencent/devops/stream/trigger/parsers/PipelineDelete.kt
@@ -80,6 +80,25 @@ class PipelineDelete @Autowired constructor(
         gitProjectConf: GitCIBasicSetting,
         gitToken: String
     ) {
+
+        // 直接删除分支,挪到前面，不需要对deleteYamlFiles获取后再做判断。
+        if (gitRequestEvent.isDeleteBranch()) {
+            val pipelines = streamPipelineBranchService.getBranchPipelines(
+                gitRequestEvent.gitProjectId,
+                gitRequestEvent.branch
+            )
+            pipelines.forEach {
+                // 这里需增加获取pipeline对应的yml路径，需要判断该文件是否在默认分支存在。
+                val gitPipelineResourceRecord = gitPipelineResourceDao.getPipelineById(
+                    dslContext =  dslContext,
+                    gitProjectId = gitRequestEvent.gitProjectId,
+                    pipelineId = it
+                )
+                delete(it, gitRequestEvent, gitPipelineResourceRecord?.filePath, gitProjectConf, gitToken)
+            }
+            return
+        }
+
         val deleteYamlFiles = when (event) {
             is GitPushEvent -> {
                 event.commits?.flatMap {
@@ -109,19 +128,6 @@ class PipelineDelete @Autowired constructor(
                 null
             }
         }
-
-        // 直接删除分支
-        if (gitRequestEvent.isDeleteBranch()) {
-            val pipelines = streamPipelineBranchService.getBranchPipelines(
-                gitRequestEvent.gitProjectId,
-                gitRequestEvent.branch
-            )
-            pipelines.forEach {
-                delete(it, gitRequestEvent, null, gitProjectConf)
-            }
-            return
-        }
-
         if (deleteYamlFiles.isNullOrEmpty()) {
             return
         }
@@ -129,7 +135,7 @@ class PipelineDelete @Autowired constructor(
         deleteYamlFiles.forEach { filePath ->
             val existPipeline = path2PipelineExists[filePath] ?: return@forEach
             val pipelineId = existPipeline.pipelineId
-            delete(pipelineId, gitRequestEvent, filePath, gitProjectConf)
+            delete(pipelineId, gitRequestEvent, filePath, gitProjectConf, gitToken)
         }
     }
 
@@ -137,7 +143,8 @@ class PipelineDelete @Autowired constructor(
         pipelineId: String,
         gitRequestEvent: GitRequestEvent,
         filePath: String?,
-        gitProjectConf: GitCIBasicSetting
+        gitProjectConf: GitCIBasicSetting,
+        gitToken: String
     ) {
         val processClient = client.get(ServicePipelineResource::class)
         // 先删除后查询的过程需要加锁
@@ -153,7 +160,11 @@ class PipelineDelete @Autowired constructor(
                 pipelineId,
                 gitRequestEvent.branch
             )
-            if (!streamPipelineBranchService.hasBranchExist(gitRequestEvent.gitProjectId, pipelineId)) {
+            val checkFileEmpty = filePath?.let { self ->
+                scmService.getFileTreeFromGitWithDefaultBranch(gitToken, gitRequestEvent.gitProjectId, self).isEmpty()
+            }
+
+            if (checkFileEmpty == true && !streamPipelineBranchService.hasBranchExist(gitRequestEvent.gitProjectId, pipelineId)) {
                 logger.info("event: ${gitRequestEvent.id} delete file: $filePath with pipeline: $pipelineId ")
                 gitPipelineResourceDao.deleteByPipelineId(dslContext, pipelineId)
                 processClient.delete(
